requirements.txt
pygame
torch
torchvision
matplotlib
ipython
pyyaml

numpy